+++ spec/controllers/doorkeeper/authorizations_controller_spec.rb
@@ -382,6 +382,34 @@ describe Doorkeeper::AuthorizationsController, type: :controller do
       end
     end
 
+    {
+      'an Integer (epoch seconds)' => ->(time) { time.to_i },
+      'a Time' => ->(time) { time.getlocal },
+      'an ActiveSupport::TimeWithZone' => ->(time) { time.in_time_zone },
+    }.each do |type_description, converter|
+      context "when auth_time_from_resource_owner returns #{type_description}" do
+        before do
+          allow(Doorkeeper::OpenidConnect.configuration)
+            .to receive(:auth_time_from_resource_owner)
+            .and_return(->(resource_owner) { converter.call(resource_owner.current_sign_in_at) })
+        end
+
+        it 'renders the authorization form if the last login is within max_age' do
+          user.update! current_sign_in_at: 5.seconds.ago
+          authorize! max_age: 10
+
+          expect_authorization_form!
+        end
+
+        it 'reauthenticates the user if the last login is older than max_age' do
+          user.update! current_sign_in_at: 15.seconds.ago
+          authorize! max_age: 10
+
+          expect(response).to redirect_to '/reauthenticate'
+        end
+      end
+    end
+
     context 'when used along with prompt=select_account' do
       it 'renders the authorization form' do
         user.update! current_sign_in_at: 5.seconds.ago